event-service/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
asyncpg
python-jose
